Werror check

+++ C++ Module 01/ex01/main.cpp
@@ -1,7 +1,7 @@
 #include "Zombie.h"
 
 int	main(void) {
-	size_t	zombie_nbr = 4;
+	int	zombie_nbr = 4;
 	Zombie *horde = zombieHorde(zombie_nbr, "Alice");
 	for (int i = 0; i < zombie_nbr; i++) {
 		horde[i].announce();
